sidebar styling and removed testpage, 3d comes from /chat now
